% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-deseqtransfrom.R
\name{DESeq2_tidiers}
\alias{DESeq2_tidiers}
\alias{tidy.DESeqTransform}
\title{Tidying methods for DESeq2 DESeqTransform objects}
\usage{
\method{tidy}{DESeqTransform}(x, colData = FALSE, ...)
}
\arguments{
\item{x}{DESeqTransform object}

\item{colData}{whether colData should be included in the tidied output
for those in the DESeqTransform object.}

\item{...}{extra arguments (not used)}
}
\value{
the result is a data frame with the columns
\item{gene}{gene ID}
\item{sample}{sample ID}
\item{log2Count}{transformed normalized counts in this gene in this sample on a log2 scale}

If \code{colData = TRUE}, it also merges this with the columns present
in \code{colData(x)}.
}
\description{
This reshapes a DESeq2 transform object (output from \code{rlog()} and
\code{varianceStabilizingTransform()} functions) into a tidy format.
}
\details{
\code{colDat=TRUE} adds covariates from colData to the data frame.
}
\note{
Unclear what the license for this code might be.  It's copied directly from https://gist.github.com/tavareshugo/3973461a7daf8a43e65e3566d5deed14
}
\examples{

# From DESeq2 documentation

if (require("DESeq2")) {
    dds = makeExampleDESeqDataSet(betaSD = 1)

    # vst transform
    vst_norm = varianceStabilizingTransformation(dds)

    # rlog transform
    rlog_norm = rlog(dds)

    tidy(rlog_norm)
    tidy(vst_norm)

    # With design included
    tidy(rlog_norm, colData=TRUE)
    tidy(vst_norm, colData=TRUE)
}

}
\author{
Hugo Tavares https://gist.github.com/tavareshugo/3973461a7daf8a43e65e3566d5deed14
}
