% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NebulaSE-methods.R
\name{gresults_names.NebulaSE}
\alias{gresults_names.NebulaSE}
\alias{gresults.NebulaSE}
\title{Return DESeq2-style \code{results}}
\usage{
\method{gresults_names}{NebulaSE}(object)

\method{gresults}{NebulaSE}(object, contrast, name, ...)
}
\arguments{
\item{object}{output from \code{\link[=run_nebula]{run_nebula()}}}

\item{contrast}{\code{numeric}, giving linear combination of coefficients}

\item{name}{a name of a term.}

\item{...}{unused}
}
\value{
\code{tibble}
}
\description{
One and only one of \code{contrast} or \code{name} must be specified.
}
\section{Functions}{
\itemize{
\item \code{gresults_names.NebulaSE}: return the names of the model coefficients.
}}

