% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{gresults}
\alias{gresults}
\alias{gresults_names}
\title{Extract results from a differential expression fit}
\usage{
gresults(object, ...)

gresults_names(object)
}
\arguments{
\item{object}{an object containing differential expression results}

\item{...}{optional parameters}
}
\value{
\code{tibble}
}
\description{
Extract results from a differential expression fit
}
\section{Functions}{
\itemize{
\item \code{gresults_names}: names of coefficients in \code{\link[=gresults]{gresults()}}
}}

