% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin-mixed-de.R
\name{has_re}
\alias{has_re}
\title{Do a set of formulaes have a random effect?}
\usage{
has_re(formulae)
}
\arguments{
\item{formulae}{vector that can be coerced to formula (e.g. \code{character()})}
}
\value{
\code{logical}
}
\description{
Do a set of formulaes have a random effect?
}
\examples{
has_re(list(~ 1 + (1 |x), "~ treatment + (1|group)"))
}
