% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/NebulaSE-methods.R
\docType{class}
\name{NebulaSE-class}
\alias{NebulaSE-class}
\alias{as-DESeqDataSet-method}
\alias{tidy.NebulaSE}
\alias{tidy,NebulaSE-method}
\title{SummarizedExperiment with Nebula results}
\usage{
as(x, 'NebulaSE')

\method{tidy}{NebulaSE}(x, ...)

\S4method{tidy}{NebulaSE}(x, ...)
}
\arguments{
\item{x}{output from \code{\link[=run_nebula]{run_nebula()}}}

\item{...}{not used}
}
\value{
\code{tibble} with columns matching \code{\link[biobroom:DESeq2_tidiers]{biobroom::tidy.DESeqResults()}}
}
\description{
A subclass that provides no additional slots,
but exists to implement polymorphic behavior of \code{tidy}, \code{results} and \code{resultsNames}
}
\section{Functions}{
\itemize{
\item \code{tidy.NebulaSE}: return tidy results

\item \code{tidy,NebulaSE-method}: S4 tidy
}}

\seealso{
\code{\link[=run_nebula]{run_nebula()}}
}
