% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{run_deseq_design}
\alias{run_deseq_design}
\title{Run DESeq2 with a specified formula, on a specified subset of data}
\usage{
run_deseq_design(dge_, formula_, subset_)
}
\arguments{
\item{dge_}{\code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet()}} object}

\item{formula_}{\code{\link[=formula]{formula()}} with a simple random intercept}

\item{subset_}{an R expression, as a \code{character}, evaluated in the context of \code{colData(dge_)} to subset the columns (samples)}
}
\value{
\code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet()}}
}
\description{
Run DESeq2 with a specified formula, on a specified subset of data
}
