% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin-mixed-de.R
\name{run_glmm_nb_design}
\alias{run_glmm_nb_design}
\title{Differential Expression with Negative Binomial Mixed Models}
\usage{
run_glmm_nb_design(
  dge_,
  formula_,
  assay_ = "counts",
  subset_,
  method = c("nebula", "glmer.nb"),
  ...
)
}
\arguments{
\item{dge_}{\code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet()}} object}

\item{formula_}{\code{\link[=formula]{formula()}} with a simple random intercept}

\item{assay_}{which assay in \code{dge_} to use}

\item{subset_}{an R expression, as a \code{character}, evaluated in the context of \code{colData(dge_)} to subset the columns (samples)}

\item{method}{\code{character}}

\item{...}{passed to \code{\link[lme4:glmer.nb]{lme4::glmer.nb()}} or \code{\link[nebula:nebula]{nebula::nebula()}}}
}
\value{
an object with a \code{tidy}, \code{results} and \code{resultNames} method
}
\description{
Differential Expression with Negative Binomial Mixed Models
}
