% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin-mixed-de.R
\name{run_nebula}
\alias{run_nebula}
\title{Run a negative binomial mixed model regression using Nebula}
\usage{
run_nebula(dge_, formula_, ...)
}
\arguments{
\item{dge_}{DESeq2 object}

\item{formula_}{formula}

\item{...}{passed to \code{\link[nebula:nebula]{nebula::nebula()}}}
}
\value{
NebulaSE-class
}
\description{
Run a negative binomial mixed model regression using Nebula
}
