% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{se_pca_plot}
\alias{se_pca_plot}
\title{Run PCA and return a plot}
\usage{
se_pca_plot(object, ntop = 500, aes_point, aes_label, print_plot = TRUE)
}
\arguments{
\item{object}{\code{SummarizedExperiment}}

\item{ntop}{use this many variable genes}

\item{aes_point}{a call to \code{aes} evaluated in the context of the colData for the \code{geom_point}, eg \code{aes(color = treatment)}.}

\item{aes_label}{a call to \code{aes} evaluated in the context of the colData for the \code{geom_textrepel}, eg \code{aes(label = sample_id)}.}

\item{print_plot}{\code{logical}}
}
\value{
list with elements \code{pca_dat}, containing \code{colData} plus PC1, PC2, \code{pca} output of \code{prcomp} and \code{plot}.
}
\description{
Run PCA and return a plot
}
